Decoupling GDI specific logic from OpenKH low level libraries
