+++ OpenKh.Tools.ImageViewer/Services/ImageFormatService.cs
@@ -7,6 +7,7 @@ using OpenKh.Common;
 using OpenKh.Imaging;
 using OpenKh.Kh2;
 using OpenKh.Kh2.Extensions;
+using OpenKh.Tools.Common.Imaging;
 
 namespace OpenKh.Tools.ImageViewer.Services
 {
